Changed the ui of card
